+++ src/aind_metadata_mapper/U19/procedures.py
