+++ tests/test_u19_etl.py
@@ -16,7 +16,7 @@ from aind_data_schema.core.procedures import (
 )
 from aind_data_schema_models.organizations import Organization
 
-from aind_metadata_mapper.U19.u19_etl import (
+from aind_metadata_mapper.U19.procedures import (
     JobSettings,
     U19Etl,
     get_dates,
@@ -66,7 +66,7 @@ class TestU19Writer(unittest.TestCase):
 
         self.assertEqual(etl1.job_settings, etl0.job_settings)
 
-    @patch("aind_metadata_mapper.U19.u19_etl.U19Etl.download_procedure_file")
+    @patch("aind_metadata_mapper.U19.procedures.U19Etl.download_procedure_file")
     def test_run_job(self, mock_download_procedure):
         """Test run_job method."""
 
@@ -80,7 +80,7 @@ class TestU19Writer(unittest.TestCase):
 
         self.assertEqual(self.example_output, actual_output)
 
-    @patch("aind_metadata_mapper.U19.u19_etl.U19Etl.download_procedure_file")
+    @patch("aind_metadata_mapper.U19.procedures.U19Etl.download_procedure_file")
     def test_extract(self, mock_download_procedure):
         """Test extract method."""
 
@@ -117,7 +117,7 @@ class TestU19Writer(unittest.TestCase):
             ),
         )
 
-    @patch("aind_metadata_mapper.U19.u19_etl.U19Etl._transform")
+    @patch("aind_metadata_mapper.U19.procedures.U19Etl._transform")
     def test_load(self, mock_transform):
         """Test load method."""
